Add architecture for test system
