+++ src/main/solution_space/code_1.py
