+++ src/main/solution_space/solution_space_handler.py
@@ -9,7 +9,7 @@ import pandas as pd
 
 from src.main.util import consts
 from src.main.util.data_util import Column
-from src.main.solution_space.code import Code
+from src.main.solution_space.code_1 import Code
 from src.main.canonicalization.consts import TREE_TYPE
 from src.main.util.log_util import log_and_raise_error
 from src.main.splitting.splitting import unpack_tests_results
